+++ src/main/java/roomescape/application/payment/toss/TossPaymentValidator.java
@@ -1,8 +1,9 @@
-package roomescape.application.payment;
+package roomescape.application.payment.toss;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.application.payment.dto.TossPaymentValidationCommand;
+import roomescape.application.payment.OrderAmountVerificationCache;
+import roomescape.application.payment.toss.dto.TossPaymentValidationCommand;
 
 @Service
 @RequiredArgsConstructor
