+++ src/main/java/roomescape/presentation/api/payment/TossPaymentController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.payment.TossPaymentValidator;
+import roomescape.application.payment.toss.TossPaymentValidator;
 import roomescape.presentation.api.payment.dto.TossPaymentValidationRequest;
 
 @RestController
