+++ src/test/java/roomescape/application/payment/CreatePaymentServiceTest.java
@@ -3,7 +3,8 @@ package roomescape.application.payment;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import roomescape.application.payment.dto.TossPaymentValidationCommand;
+import roomescape.application.payment.toss.dto.TossPaymentValidationCommand;
+import roomescape.application.payment.toss.TossPaymentValidator;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
